refactor :: Winners 테스트 리팩토링
