+++ src/main/java/domain/Winners.java
@@ -3,6 +3,7 @@ package domain;
 import static util.Validator.validateArgument;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.StringJoiner;
 
 public class Winners {
@@ -19,9 +20,9 @@ public class Winners {
 
   @Override
   public String toString() {
-      StringJoiner stringJoiner = new StringJoiner(WINNERS_DELIMITER);
-      winners.forEach(stringJoiner::add);
-      return stringJoiner.toString();
+    StringJoiner stringJoiner = new StringJoiner(WINNERS_DELIMITER);
+    winners.forEach(stringJoiner::add);
+    return stringJoiner.toString();
   }
 
   public void validate(List<String> winners) {
@@ -30,5 +31,22 @@ public class Winners {
         (arg) -> arg != null && !arg.isEmpty(),
         MESSAGE_FOR_INVALIDATE_WINNERS
     );
+  }
+
+  @Override
+  public boolean equals(Object o) {
+    if (this == o) {
+      return true;
+    }
+    if (o == null || getClass() != o.getClass()) {
+      return false;
+    }
+    Winners otherWinners = (Winners) o;
+    return winners.containsAll(otherWinners.winners) && otherWinners.winners.containsAll(winners);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(winners);
   }
 }
